Revert "refactor(proposal): Rename for review proposal request."
This reverts commit e10336dcf86c6bcda3bc9c2aad586c4234140184.
