+++ app/Http/Controllers/Quotes/Proposals/CreateQuoteProposalController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Quotes\Proposals;
 
+use App\Constants\QuoteProposalTypeConstant;
 use App\Constants\QuoteTypesConstant;
 use App\Constants\UserTypesAbilitiesConstant;
 use App\Http\Controllers\Controller;
